third_party/WebKit/LayoutTests/external/wpt/encoding/legacy-mb-japanese/iso-2022-jp/iso2022jp-encode-href-ascii-expected.txt

This is a testharness.js-based test.
FAIL %65 assert_equals: expected "%65%3B" but got "%65"
FAIL %1B assert_equals: expected "%64%3B" but got ""
Harness: the test ran to completion.
